+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * `peek_var()` error now generates hyperlink to docs with recent RStudio (#289).
 
+* `all_of()` (like `any_of()`) now fails when used outside of a
+  tidyselect context (#269).
+
 * `num_range()` gains a `suffix` argument (#229).
 
 * `eval_select()` always returns a named vector, even when renaming is not